% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_weights}
\alias{plot_weights}
\title{Spatially plot the confident weights for each cell type}
\usage{
plot_weights(
  cell_type_names,
  puck,
  resultsdir,
  weights,
  minUMI = 100,
  maxUMI = 2e+05,
  min_val = 0,
  max_val = 1,
  counter_barcodes = NULL,
  my_pal = NULL,
  xlim = NULL,
  ylim = NULL,
  size = 0.15,
  alpha = 1,
  small_point = F
)
}
\arguments{
\item{cell_type_names}{list of cell type names}

\item{puck}{an object of type \linkS4class{SpatialRNA}}

\item{resultsdir}{output directory}

\item{weights}{a dataframe of RCTD output weights (stored in the RCTD object)}

\item{minUMI}{numeric, minimum value for total UMIs to filter pixels}

\item{maxUMI}{numeric, maximum value for total UMIs to filter pixels}

\item{min_val}{numeric, minimum value for the range of plot as a numeric list}

\item{max_val}{numeric, maximum value for the range of plot as a numeric list}

\item{xlim}{(optional) minimum and maximum value for x coordinate as a numeric list}

\item{ylim}{(optional) minimum and maximum value for y coordinate as a numeric list}

\item{size}{numeric size of points}
}
\description{
Plots the confident weights for each cell type as in full_mode. Saves as 'cell_type_weights.pdf'
}
